+++ Gulpfile.js
@@ -1,7 +1,7 @@
 //
 // Gulpfile for irc-hybrid-client
 //
-const {src, dest, watch, series, parallel} = require('gulp');
+const { src, dest, watch, series, parallel } = require('gulp');
 const del = require('del');
 const concat = require('gulp-concat');
 const insert = require('gulp-insert');
@@ -9,9 +9,10 @@ const minify = require('gulp-minify');
 const htmlmin = require('gulp-htmlmin');
 const cleancss = require('gulp-clean-css');
 const replace = require('gulp-replace');
+const rename = require('gulp-rename');
 const fs = require('fs');
 
-// For configuratino see: https://www.npmjs.com/package/terser#compress-options
+// For configuration see: https://www.npmjs.com/package/terser#compress-options
 
 // To disable length limit, set maxLineLength = false
 const maxLineLength = 200;
@@ -51,145 +52,262 @@ const minifyCssOptions = {
 };
 
 //
-// Clean secure-minify folder
+// Clean build folder
 //
-const clean = function() {
+const cleanDev = function () {
   return del(
     [
-      '../irc-hybrid-client/secure-minify'
+      '../irc-hybrid-client/build-dev'
     ],
-    {dryRun: false, force: true});
+    {
+      dryRun: false,
+      force: true
+    });
 };
 
-// for /irc/webclient.html
-const htmlMinify = function() {
-  return src('../irc-hybrid-client/secure/webclient.html')
-    // these files have been bundled into one file, so they are being removed
-    .pipe(replace('<script src="./js/webclient02.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient03.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient04.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient05.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient06.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient07.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient08.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient09.js" defer></script>', ''))
-    .pipe(replace('<script src="./js/webclient10.js" defer></script>', ''))
-    .pipe(htmlmin(htmlMinifyOptions))
-    .pipe(dest('../irc-hybrid-client/secure-minify'));
+const cleanProd = function () {
+  return del(
+    [
+      '../irc-hybrid-client/build-prod'
+    ],
+    {
+      dryRun: false,
+      force: true
+    });
 };
-// For /irc/serverlist.html
-const htmlMinify2 = function() {
-  return src('../irc-hybrid-client/secure/serverlist.html')
-    .pipe(htmlmin(htmlMinifyOptions))
-    .pipe(dest('../irc-hybrid-client/secure-minify'));
+
+const placeholderDev = function () {
+  return src('../irc-hybrid-client/source-files/html/build-placeholder.html')
+  .pipe(rename('webclient.html'))
+  .pipe(dest('../irc-hybrid-client/build-dev'));
+}
+
+const placeholderProd = function () {
+  return src('../irc-hybrid-client/source-files/html/build-placeholder.html')
+  .pipe(rename('webclient.html'))
+  .pipe(dest('../irc-hybrid-client/build-prod'));
+}
+
+// -----------------
+//   page01.html (previously index.html)
+// -----------------
+const htmlWebclientDev = function () {
+  const templates = '\n' +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/websocket-panel.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/nav-menu.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/activity-spinner.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/hamburger-icon.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/header-bar.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/error-panel.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/notice-panel.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/manage-channels-panel.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/channel-panel.html', 'utf8') +
+    fs.readFileSync('../irc-hybrid-client/source-files/web-components/debug-panel.html', 'utf8');
+  const jsFilenames = '\n' +
+    '<script src="./js/_beforeLoad.js" defer></script>\n' +
+    '<script src="./js/temp-placeholder.js" defer></script>\n' +
+    '<script src="./js/glob-vars.js" defer></script>\n' +
+    '<script src="./js/display-utils.js" defer></script>\n' +
+    '<script src="./js/user-info.js" defer></script>\n' +
+    '<script src="./js/websocket-panel.js" defer></script>\n' +
+    '<script src="./js/nav-menu.js" defer></script>\n' +
+    '<script src="./js/activity-spinner.js" defer></script>\n' +
+    '<script src="./js/hamburger-icon.js" defer></script>\n' +
+    '<script src="./js/header-bar.js" defer></script>\n' +
+    '<script src="./js/error-panel.js" defer></script>\n' +
+    '<script src="./js/notice-panel.js" defer></script>\n' +
+    '<script src="./js/manage-channels-panel.js" defer></script>\n' +
+    '<script src="./js/channel-panel.js" defer></script>\n' +
+    '<script src="./js/debug-panel.js" defer></script>\n' +
+    '<script src="./js/_afterLoad.js" defer></script>\n';
+  const now = new Date();
+  const compileDate = 'Build timestamp: ' + now.toGMTString();
+  return src('../irc-hybrid-client/source-files/html/_index.html')
+    .pipe(rename('webclient.html'))
+    .pipe(replace('<!-- HTML TEMPLATES -->', templates))
+    .pipe(replace('<!-- COMPILE-DATE -->', compileDate))
+    .pipe(replace('<script src="./js/webclient.js" defer></script>', jsFilenames))
+    .pipe(dest('../irc-hybrid-client/build-dev'));
 };
-// For js/webclient.js
-const jsMinify = function () {
-  const license = '/*\n' +
-    fs.readFileSync('LICENSE', 'utf8') +
-    '*/\n\n';
+
+const htmlWebclientProd = function () {
+  const templates = '' +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/websocket-panel.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/nav-menu.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/activity-spinner.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/hamburger-icon.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/header-bar.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/error-panel.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/notice-panel.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/manage-channels-panel.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/channel-panel.html', 'utf8') +
+  fs.readFileSync('../irc-hybrid-client/source-files/web-components/debug-panel.html', 'utf8');
+const now = new Date();
+  const compileDate = 'Build timestamp: ' + now.toGMTString();
+  return src('../irc-hybrid-client/source-files/html/_index.html')
+    .pipe(rename('webclient.html'))
+    .pipe(replace('<!-- HTML TEMPLATES -->', templates))
+    .pipe(replace('<!-- COMPILE-DATE -->', compileDate))
+    .pipe(htmlmin(htmlMinifyOptions))
+    .pipe(dest('../irc-hybrid-client/build-prod'));
+};
+const jsWebclientDev = function () {
+  return src(
+    [
+      '../irc-hybrid-client/source-files/js/_beforeLoad.js',
+      '../irc-hybrid-client/source-files/web-components/temp-placeholder.js',
+      '../irc-hybrid-client/source-files/web-components/glob-vars.js',
+      '../irc-hybrid-client/source-files/web-components/display-utils.js',
+      '../irc-hybrid-client/source-files/web-components/user-info.js',
+      '../irc-hybrid-client/source-files/web-components/websocket-panel.js',
+      '../irc-hybrid-client/source-files/web-components/nav-menu.js',
+      '../irc-hybrid-client/source-files/web-components/activity-spinner.js',
+      '../irc-hybrid-client/source-files/web-components/hamburger-icon.js',
+      '../irc-hybrid-client/source-files/web-components/header-bar.js',
+      '../irc-hybrid-client/source-files/web-components/error-panel.js',
+      '../irc-hybrid-client/source-files/web-components/notice-panel.js',
+      '../irc-hybrid-client/source-files/web-components/manage-channels-panel.js',
+      '../irc-hybrid-client/source-files/web-components/channel-panel.js',
+      '../irc-hybrid-client/source-files/web-components/debug-panel.js',
+      '../irc-hybrid-client/source-files/js/_afterLoad.js'
+    ])
+    .pipe(dest('../irc-hybrid-client/build-dev/js'));
+};
+
+const jsWebclientProd = function () {
+  const license = '/*\n' + fs.readFileSync('LICENSE', 'utf8') + '*/\n\n';
   const jsStrict = '\'use strict\';\n\n';
   return src(
     [
-      '../irc-hybrid-client/secure/js/webclient.js',
-      '../irc-hybrid-client/secure/js/webclient02.js',
-      '../irc-hybrid-client/secure/js/webclient03.js',
-      '../irc-hybrid-client/secure/js/webclient04.js',
-      '../irc-hybrid-client/secure/js/webclient05.js',
-      '../irc-hybrid-client/secure/js/webclient06.js',
-      '../irc-hybrid-client/secure/js/webclient07.js',
-      '../irc-hybrid-client/secure/js/webclient08.js',
-      '../irc-hybrid-client/secure/js/webclient09.js',
-      '../irc-hybrid-client/secure/js/webclient10.js'
+      '../irc-hybrid-client/source-files/js/_beforeLoad.js',
+      '../irc-hybrid-client/source-files/web-components/temp-placeholder.js',
+      '../irc-hybrid-client/source-files/web-components/glob-vars.js',
+      '../irc-hybrid-client/source-files/web-components/display-utils.js',
+      '../irc-hybrid-client/source-files/web-components/user-info.js',
+      '../irc-hybrid-client/source-files/web-components/websocket-panel.js',
+      '../irc-hybrid-client/source-files/web-components/nav-menu.js',
+      '../irc-hybrid-client/source-files/web-components/activity-spinner.js',
+      '../irc-hybrid-client/source-files/web-components/hamburger-icon.js',
+      '../irc-hybrid-client/source-files/web-components/header-bar.js',
+      '../irc-hybrid-client/source-files/web-components/error-panel.js',
+      '../irc-hybrid-client/source-files/web-components/notice-panel.js',
+      '../irc-hybrid-client/source-files/web-components/manage-channels-panel.js',
+      '../irc-hybrid-client/source-files/web-components/channel-panel.js',
+      '../irc-hybrid-client/source-files/web-components/debug-panel.js',
+      '../irc-hybrid-client/source-files/js/_afterLoad.js'
     ])
     .pipe(replace('\'use strict\';\n', ''))
     .pipe(concat('webclient.js'))
     .pipe(minify(jsMinifyOptions))
     .pipe(insert.prepend(jsStrict))
     .pipe(insert.prepend(license))
-    .pipe(dest('../irc-hybrid-client/secure-minify/js'));
-};
-// For js/serverlist.js
-const jsMinify2 = function () {
-  const license = '/*\n' +
-    fs.readFileSync('LICENSE', 'utf8') +
-    '*/\n\n';
-  const jsStrict = '\'use strict\';\n\n';
-  return src('../irc-hybrid-client/secure/js/serverlist.js')
-    .pipe(replace('\'use strict\';\n', ''))
-    .pipe(minify(jsMinifyOptions))
-    .pipe(insert.prepend(jsStrict))
-    .pipe(insert.prepend(license))
-    .pipe(dest('../irc-hybrid-client/secure-minify/js'));
+    .pipe(dest('../irc-hybrid-client/build-prod/js'));
 };
 
-
-// -----------
-//    CSS
-// -----------
-// for css/styles.css
-const cssMinify = function() {
+const cssWebclientDev = function () {
   return src(
     [
-      '../irc-hybrid-client/secure/css/styles.css'
+      '../irc-hybrid-client/source-files/css/_global.css',
+      '../irc-hybrid-client/source-files/css/_styles-old.css',
+      '../irc-hybrid-client/source-files/web-components/websocket-panel.css',
+      '../irc-hybrid-client/source-files/web-components/nav-menu.css',
+      '../irc-hybrid-client/source-files/web-components/activity-spinner.css',
+      '../irc-hybrid-client/source-files/web-components/hamburger-icon.css',
+      '../irc-hybrid-client/source-files/web-components/header-bar.css',
+      '../irc-hybrid-client/source-files/web-components/error-panel.css',
+      '../irc-hybrid-client/source-files/web-components/notice-panel.css',
+      '../irc-hybrid-client/source-files/web-components/manage-channels-panel.css',
+      '../irc-hybrid-client/source-files/web-components/debug-panel.css'
     ])
     .pipe(concat('styles.css'))
-    .pipe(cleancss(minifyCssOptions))
-    .pipe(dest('../irc-hybrid-client/secure-minify/css'));
+    .pipe(dest('../irc-hybrid-client/build-dev/css'));
 };
-// for css/serverlist.css
-const cssMinify2 = function() {
-  return src('../irc-hybrid-client/secure/css/serverlist.css')
-    .pipe(cleancss(minifyCssOptions))
-    .pipe(dest('../irc-hybrid-client/secure-minify/css'));
-};
+
+const cssWebclientProd = function () {
+  return src('../irc-hybrid-client/build-dev/css/styles.css')
+  .pipe(cleancss(minifyCssOptions))
+  .pipe(dest('../irc-hybrid-client/build-prod/css'));
+}
+
 
 // ------------------------------
 // copy sound files
 // ------------------------------
-const soundsCopy = function() {
-  return src('../irc-hybrid-client/secure/sounds/short-beep*')
-    .pipe(dest('../irc-hybrid-client/secure-minify/sounds'));
+const soundsCopyDev = function() {
+  return src('../irc-hybrid-client/source-files/sounds/short-beep*')
+    .pipe(dest('../irc-hybrid-client/build-dev/sounds'));
 };
-
-// ---------------------------------------------
-// Check for existence of irc-hybrid-client
-// ---------------------------------------------
-const verifyFoldersExist = function (cb) {
-  if (!fs.existsSync('../irc-hybrid-client')) {
-    throw new Error('Folder ../irc-hybrid-client/ does not exist');
-  }
-  if (!fs.existsSync('../irc-hybrid-client/secure')) {
-    throw new Error('Folder ../irc-hybrid-client/secure/ does not exist');
-  }
-  cb();
+const soundsCopyProd = function() {
+  return src('../irc-hybrid-client/source-files/sounds/short-beep*')
+    .pipe(dest('../irc-hybrid-client/build-prod/sounds'));
 };
 
 //
-// default using 'gulp' command
+// Process Production build folder
 //
-const defaultTask = function (cb) {
-  console.log('\n\nTo Minify HTML file use: npx gulp minify\n\n');
-  cb();
-};
-
+const buildProd = series(
+  htmlWebclientProd,
+  jsWebclientProd,
+  cssWebclientDev,
+  cssWebclientProd,
+  soundsCopyProd
+);
 //
-// Process Production secure-minify folder
+// Process HTML files
 //
-const minifyProd = series(
-  // For /irc/webclient.html
-  htmlMinify,
-  jsMinify,
-  cssMinify,
-  // For /irc/serverlist.html
-  htmlMinify2,
-  jsMinify2,
-  cssMinify2,
-  // other files
-  soundsCopy
+const buildDev = series(
+  htmlWebclientDev,
+  jsWebclientDev,
+  cssWebclientDev,
+  soundsCopyDev
 );
 
-exports.default = defaultTask;
-exports.clean = series(verifyFoldersExist, clean);
-exports.minify = series(verifyFoldersExist, clean, minifyProd);
-exports.dist = series(verifyFoldersExist, clean, minifyProd);
+//
+// dynamically rebuild index.html as needed
+//
+const watchHtml = function () {
+  console.log('watchHtml, waiting for html changes to rebuild index.html');
+  return watch(
+    [
+      '../irc-hybrid-client/source-files/html/*.html',
+      '../irc-hybrid-client/source-files/css/*.css',
+      '../irc-hybrid-client/source-files/js/*.js',
+      '../irc-hybrid-client/source-files/web-components/*.html',
+      '../irc-hybrid-client/source-files/web-components/*.js',
+      '../irc-hybrid-client/source-files/web-components/*.css'
+    ],
+    {
+      ignoreInitial: true
+    },
+    buildDev);
+};
+
+exports.cleanDev = series(
+  cleanDev,
+  placeholderDev,
+);
+exports.cleanDist = series(
+  cleanProd,
+  placeholderProd
+);
+exports.cleanAll = series(
+  cleanDev,
+  placeholderDev,
+  cleanProd,
+  placeholderProd
+);
+exports.default = series(
+  cleanDev,
+  buildDev,
+  watchHtml
+);
+exports.dev = series(
+  cleanDev,
+  buildDev
+);
+exports.dist = series(
+  cleanDev,
+  cleanProd,
+  buildDev,
+  buildProd,
+);